+++ packages/emulators/standalone/mupen64plus-sa/mupen64plus-sa-video-glide64mk2/package.mk
@@ -1,5 +1,7 @@
 # SPDX-License-Identifier: GPL-2.0-or-later
 # Copyright (C) 2019-present Shanti Gilbert (https://github.com/shantigilbert)
+# Copyright (C) 2023 Nicholas Ricciuti (rishooty@gmail.com)
+# Copyright (C) 2023 Fewtarius (fewtarius@jelos.org)
 
 PKG_NAME="mupen64plus-sa-video-glide64mk2"
 PKG_VERSION="d900f2191575e01eb846a1009be71cbc1b413dba"
@@ -12,43 +14,48 @@ PKG_LONGDESC="Mupen64Plus Standalone Glide64 Video Driver"
 PKG_TOOLCHAIN="manual"
 
 case ${DEVICE} in
-  AMD64|RK3588|S922X)
-    PKG_DEPENDS_TARGET="mupen64plus-sa-simplecore"
+  AMD64|RK3588|S922X|RK3399)
+    PKG_DEPENDS_TARGET+=" mupen64plus-sa-simplecore"
   ;;
 esac
 
-if [ ! "${OPENGL}" = "no" ]; then
-  PKG_DEPENDS_TARGET+=" ${OPENGL} glu libglvnd"
-fi
-
-if [ "${OPENGLES_SUPPORT}" = yes ]; then
-  PKG_DEPENDS_TARGET+=" ${OPENGLES}"
-fi
+case ${DEVICE} in
+  AMD64)
+    PKG_DEPENDS_TARGET+=" ${OPENGL} glu libglvnd"
+    export USE_GLES=0
+  ;;
+  *)
+    PKG_DEPENDS_TARGET+=" ${OPENGLES}"
+    export USE_GLES=1
+  ;;
+esac
 
 make_target() {
-  case ${ARCH} in
-    arm|aarch64)
-      export HOST_CPU=aarch64
-      BINUTILS="$(get_build_dir binutils)/.aarch64-libreelec-linux-gnueabi"
-      export USE_GLES=1
-    ;;
-    x86_64)
-      export HOST_CPU=x86_64
-      export USE_GLES=0
-    ;;
-  esac
-  export APIDIR=${SYSROOT_PREFIX}/usr/local/include/mupen64plus
-  export SDL_CFLAGS="-I${SYSROOT_PREFIX}/usr/include/SDL2 -pthread"
+
+  export HOST_CPU=${TARGET_ARCH} \
+         NEW_DYNAREC=1 \
+         VFP_HARD=1 \
+         V=1 \
+         VC=0 \
+         OSD=0
+
+  export BINUTILS="$(get_build_dir binutils)/.${TARGET_NAME}"
+  export APIDIR=$(get_build_dir mupen64plus-sa-core)/src/api
+  export SDL_CFLAGS="-I${SYSROOT_PREFIX}/usr/include/SDL2 -pthread -D_REENTRANT"
   export SDL_LDLIBS="-lSDL2_net -lSDL2"
   export CROSS_COMPILE="${TARGET_PREFIX}"
-  export V=1
-  export VC=0
+
   make -C projects/unix clean
   make -C projects/unix all ${PKG_MAKE_OPTS_TARGET}
   cp ${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2.so ${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2-base.so
-  export APIDIR=${SYSROOT_PREFIX}/usr/local/include/simple64
-  make -C projects/unix all ${PKG_MAKE_OPTS_TARGET}
-  cp ${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2.so ${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2-simple.so
+
+  case ${DEVICE} in
+    AMD64|RK3588|S922X|RK3399)
+      export APIDIR=$(get_build_dir mupen64plus-sa-simplecore)/src/api
+      make -C projects/unix all ${PKG_MAKE_OPTS_TARGET}
+      cp ${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2.so ${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2-simple.so
+    ;;
+  esac
 }
 
 makeinstall_target() {
@@ -58,11 +65,14 @@ makeinstall_target() {
   UPLUGINDIR=${ULIBDIR}/mupen64plus
   mkdir -p ${UPLUGINDIR}
   cp ${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2-base.so ${UPLUGINDIR}/mupen64plus-video-glide64mk2.so
-  #${STRIP} ${UPLUGINDIR}/mupen64plus-video-glide64mk2.so
   chmod 0644 ${UPLUGINDIR}/mupen64plus-video-glide64mk2.so
-  cp ${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2-simple.so ${UPLUGINDIR}
-  #${STRIP} ${UPLUGINDIR}/mupen64plus-video-glide64mk2-simple.so
-  chmod 0644 ${UPLUGINDIR}/mupen64plus-video-glide64mk2-simple.so
+
+  if [ -e "${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2-simple.so" ]
+  then
+    cp ${PKG_BUILD}/projects/unix/mupen64plus-video-glide64mk2-simple.so ${UPLUGINDIR}
+    chmod 0644 ${UPLUGINDIR}/mupen64plus-video-glide64mk2-simple.so
+  fi
+
   mkdir -p ${USHAREDIR}
   cp ${PKG_BUILD}/data/Glide64mk2.ini ${USHAREDIR}
   chmod 0644 ${USHAREDIR}/Glide64mk2.ini
